Extract GenerateInstanceProperties from PreDeployInstance
