+++ java/configuration/src/test/java/sleeper/configuration/properties/table/TablePropertiesTestHelper.java
@@ -41,13 +41,6 @@ public class TablePropertiesTestHelper {
     }
 
     public static TableProperties createTestTableProperties(
-            InstanceProperties instanceProperties, Schema schema, String tableName, AmazonS3 s3) {
-        TableProperties tableProperties = createTestTableProperties(instanceProperties, schema, tableName);
-        save(tableProperties, s3);
-        return tableProperties;
-    }
-
-    public static TableProperties createTestTableProperties(
             InstanceProperties instanceProperties, Schema schema, AmazonS3 s3, Consumer<TableProperties> tableConfig) {
         TableProperties tableProperties = createTestTableProperties(instanceProperties, schema);
         tableConfig.accept(tableProperties);
